Rename: 폴더명 AuthInputBox -> AuthInput 수정
