+++ src/components/Auth/AuthInput/ValidifyCheck.tsx
