+++ src/components/Auth/index.ts
@@ -1,7 +1,7 @@
-import AuthInput from './AuthInputBox/AuthInput';
-import AuthEmailInputBox from './AuthInputBox/AuthEmailInputBox';
-import AuthPwInputBox from './AuthInputBox/AuthPwInputBox';
-import AuthPwCheckInputBox from './AuthInputBox/AuthPwCheckInputBox';
+import AuthInput from './AuthInput/AuthInput';
+import AuthEmailInputBox from './AuthInput/AuthEmailInputBox';
+import AuthPwInputBox from './AuthInput/AuthPwInputBox';
+import AuthPwCheckInputBox from './AuthInput/AuthPwCheckInputBox';
 import ErrorMessage from './ErrorMessage';
 import KakaoLoginButton from './KakaoLoginButton';
 import SignupButton from './SignupButton';
